@@ -15,14 +15,14 @@ class CreateNovaMediaLibraryTable extends Migration
     {
         Schema::create('nova_media_library', function (Blueprint $table) {
             $table->increments('id');
-	        $table->string('title')->index()->nullable();
-	        $table->timestamp('created')->index()->useCurrent();
-	        $table->string('type')->index();
-	        $table->string('folder')->index();
-	        $table->string('name');
-	        $table->boolean('private')->default(0);
-	        $table->boolean('lp')->default(0);
-	        $table->text('options')->nullable();
+            $table->string('title')->index()->nullable();
+            $table->timestamp('created')->index()->useCurrent();
+            $table->string('type')->index();
+            $table->string('folder')->index();
+            $table->string('name');
+            $table->boolean('private')->default(0);
+            $table->boolean('lp')->default(0);
+            $table->text('options')->nullable();
         });
     }
 
