@@ -1,31 +1,26 @@
 <template>
-  <div class="flex border-b border-40">
+    <div class="border-40 flex items-center border-b" :class="elementSize">
+        <slot>
+            <div class="text-80 nml-w-1/4 py-4 font-normal">
+                {{ field.indexName }}
+            </div>
+        </slot>
 
-    <slot>
-      <div class="w-1/4 py-4 font-normal text-80">
-        {{ field.indexName }}
-      </div>
-    </slot>
+        <slot name="value">
+            <div class="nml-field-form nml-w-3/4 py-4">
+                <template v-if="!field.value">—</template>
 
-    <slot name="value">
-      <div class="nml-field-form w-3/4 py-4">
+                <span class="dim text-primary cursor-pointer font-bold" v-else-if="isHidden" @click="isHidden = false">{{
+                    __('Show')
+                }}</span>
 
-        <template v-if="!field.value">—</template>
-
-        <span class="cursor-pointer dim text-primary font-bold"
-              v-else-if="isHidden"
-              @click="isHidden = false"
-        >{{ __('Show') }}</span>
-
-        <template v-else-if="!field.nmlJsCallback && !field.nmlTrix">
-          <nmlArray v-if="field.nmlArray" :field="field" />
-          <nmlFile v-else :field="field" />
-        </template>
-
-      </div>
-    </slot>
-
-  </div>
+                <template v-else-if="!field.nmlJsCallback && !field.nmlTrix">
+                    <nmlArray v-if="field.nmlArray" :field="field" />
+                    <nmlFile v-else :field="field" />
+                </template>
+            </div>
+        </slot>
+    </div>
 </template>
 
 <script src="./script.js"></script>
