@@ -1,23 +1,24 @@
 <template>
-  <default-field :field="field" :errors="errors" :fullWidthContent="true">
-    <template slot="field">
-      <div class="nml-field-form">
-
-        <span class="cursor-pointer dim inline-block text-primary font-bold pt-2 leading-tight"
-              v-if="isHidden && field.value"
-              @click="isHidden = false"
-        >{{ __('Show') }}</span>
-
-        <template v-else>
-          <nmlCallback v-if="field.nmlJsCallback" :field="field" />
-          <nmlTrix v-else-if="field.nmlTrix" :field="field" />
-          <nmlArray v-else-if="field.nmlArray" :field="field" :handler="handleChange" />
-          <nmlFile v-else :field="field" :handler="handleChange" />
+    <DefaultField @click="display()" :field="currentField" :errors="errors" :show-help-text="showHelpText" :fullWidthContent="true">
+        <template v-slot:field>
+            <div class="nml-field-form">
+                <span
+                    class="dim text-primary inline-block cursor-pointer pt-2 font-bold leading-tight"
+                    v-if="isHidden && field.value"
+                    @click="isHidden = false"
+                    >{{ __('Show') }}</span
+                >
+                
+                <template v-else
+                    >
+                    <nmlCallback v-if="field.nmlJsCallback" :field="currentField" />
+                    <nmlTrix v-else-if="field.nmlTrix" :field="currentField" />
+                    <nmlArray v-else-if="field.nmlArray" :field="currentField" :handler="handleChange" />
+                    <nmlFile v-else :field="currentField" :handler="handleChange" :isFormField="isFormField"/>
+                </template>
+            </div>
         </template>
-
-      </div>
-    </template>
-  </default-field>
+    </DefaultField>
 </template>
 
 <script src="./script.js"></script>
