@@ -1,12 +1,10 @@
 <template>
   <div>
-
     <div class="card border border-lg border-50 max-w-xs p-8 text-center cursor-pointer max-w-xs"
-         v-if="!item && isForm"
+         v-if="!item"
          @click="popup = true">
       {{ __('Select File') }}
     </div>
-
     <a v-else-if="item" :href="item.url" target="_blank" class="no-underline">
       <img class="block rounded-lg shadow-md max-w-xs"
            v-if="`image` === mime(item)"
@@ -26,18 +24,18 @@
     </a>
 
 
-    <div class="mt-4" v-if="isForm && item">
-      <a class="cursor-pointer dim inline-block text-primary font-bold" @click="popup = true">
+    <div class="nml-mt-4" v-if="isForm && item">
+      <a class="cursor-pointer dim inline-block text-white bg-primary-500 hover:bg-primary-700 py-2 px-4 rounded mr-autofont-bold" @click="popup = true">
         {{ __('Media Library') }}
       </a>
 
-      <a class="cursor-pointer dim inline-block text-danger font-bold ml-8" @click="changeFile(null)">
+      <a class="cursor-pointer dim inline-block text-white bg-red-500 hover:bg-red-700 py-2 nml-ml-4 px-4 rounded mr-autofont-bold" @click="changeFile(null)">
         {{ __('Clear') }}
       </a>
     </div>
 
 
-    <transition name="fade" mode="out-in">
+    <transition name="bounce">
       <Library v-if="popup" :field="field" />
     </transition>
 
@@ -46,3 +44,23 @@
 </template>
 
 <script src="./script.js"></script>
+
+<style>
+.bounce-enter-active {
+  animation: bounce-in 0.5s;
+}
+.bounce-leave-active {
+  animation: bounce-in 0.5s reverse;
+}
+@keyframes bounce-in {
+  0% {
+    transform: scale(0);
+  }
+  50% {
+    transform: scale(1.25);
+  }
+  100% {
+    transform: scale(1);
+  }
+}
+</style>
