@@ -1,32 +1,37 @@
-Nova.booting((Vue, router, store) => {
-  Vue.component('index-media-library-field', require('./field/Index/'));
-  Vue.component('detail-media-library-field', require('./field/Detail/'));
-  Vue.component('form-media-library-field', require('./field/Form/'));
+import { VueClipboard } from "@soerenmartius/vue3-clipboard";
+import Toasted from "toastedjs";
+let toasted = new Toasted({
+  theme: "nova",
+  position: "bottom-right",
+  duration: 6000,
+});
+Nova.booting((Vue) => {
+  Vue.use(VueClipboard);
+  Vue.component("index-media-library-field", require("./field/Index/").default);
+  Vue.component(
+    "detail-media-library-field",
+    require("./field/Detail/").default
+  );
+  Vue.component("form-media-library-field", require("./field/Form/").default);
+  Nova.inertia("NovaMediaLibrary", require("./tool/").default);
 
-  router.addRoutes([
-    {
-      name: 'nova-media-library',
-      path: '/media-library',
-      component: require('./tool/'),
-    },
-  ]);
-
-  window.nmlToastHook = e => {
-    if ( 422 === e.response.status && e.response.data.message )
-      Vue.prototype.$toasted.show(e.response.data.message, { type: 'error' })
+  window.nmlToastHook = (e) => {
+    if (422 === e.response.status && e.response.data.message)
+      toasted.show(e.response.data.message, { type: "error" });
   };
 });
 
-
-if ('object' === typeof Nova.config.novaMediaLibrary) {
-  if (Nova.config.novaMediaLibrary.store === 'folders') {
-    Nova.request().get('/nova-vendor/nova-media-library/folders').then(r => {
-      Object.assign(Nova.config.novaMediaLibrary, { folders: r.data })
-    })
-  }
-  if ('object' === typeof Nova.config.novaMediaLibrary.lang) {
-    Object.assign(Nova.config.translations, Nova.config.novaMediaLibrary.lang)
-  }
+// Dark mode fix
+if (localStorage.novaTheme === 'dark' || (!('novaTheme' in localStorage) && window.matchMedia('(prefers-color-scheme: dark)').matches)) {
+  document.documentElement.classList.add('nml-dark')
+} else {
+  document.documentElement.classList.remove('nml-dark')
 }
+// if ("object" === typeof window.Nova.config('novaMediaLibrary')) {
+//   if ("object" === typeof window.Nova.config('novaMediaLibrary').lang) {
+//     Object.assign(Nova.config.translations, window.Nova.config('novaMediaLibrary').lang);
+//   }
+// }
 
+//alert(window.window.Nova.config('novaMediaLibrary'));
 //document.querySelector('meta[name="viewport"]').setAttribute('content', 'width=device-width, initial-scale=1.0, user-scalable=yes');
