@@ -4,25 +4,23 @@
     <checkbox-with-label
       class="cursor-pointer mr-4"
       v-if="!$parent.field || $parent.isArray"
-      :checked="$parent.bulk.enable"
       @input="changeBulk"
-      @change="changeBulk"
     >{{ __('Bulk Select') }}</checkbox-with-label>
 
-    <button class="btn-default bg-primary text-white cursor-pointer shadow-md max-w-full mr-4"
+    <button class="bg-primary-500 hover:bg-primary-700 text-white font-bold py-1 px-2 rounded bg-primary cursor-pointer shadow-md max-w-full mr-4"
             type="button"
             @click="bulkAll"
             v-if="$parent.bulk.enable"
     >{{ __('Select All') }}</button>
 
     <template v-if="$parent.bulkLen()">
-      <button class="btn-default bg-danger text-white cursor-pointer shadow-md max-w-full mr-4"
+      <button class="bg-red-500 hover:bg-red-700 text-white font-bold py-1 px-2 rounded bg-danger cursor-pointer shadow-md max-w-full mr-4"
               type="button"
               @click="$parent.deleteFiles(Object.keys($parent.bulk.ids))">
         {{ __('Delete') }} ({{ $parent.bulkLen() }})
       </button>
 
-      <button class="btn-default bg-success text-white cursor-pointer shadow-md max-w-full mr-4"
+      <button class="bg-primary-500 hover:bg-primary-700 text-white font-bold py-1 px-2 rounded bg-success cursor-pointer shadow-md max-w-full mr-4"
               v-if="$parent.field && $parent.bulk.enable && $parent.bulkLen()"
               type="button"
               @click="pushFiles">
@@ -30,7 +28,7 @@
       </button>
     </template>
 
-    <label class="btn-default bg-90 text-white text-center cursor-pointer shadow-md max-w-full ml-auto">
+    <label class="bg-primary-500 hover:bg-primary-700 text-white font-bold py-1 px-2 rounded bg-90 text-center cursor-pointer shadow-md max-w-full ml-auto">
       <input
         id="nml_upload"
         class="form-file-input"
@@ -44,5 +42,7 @@
 
   </div>
 </template>
+
+
 
 <script src="./script.js"></script>
