@@ -1,3 +1,10 @@
+import Cropper from 'cropperjs';
+import Toasted from 'toastedjs'
+let toasted = new Toasted({
+  theme: 'nova',
+  position: 'bottom-right',
+  duration: 6000,
+})
 export default {
   props: {
     type: { type: String, default: 'folder' },
@@ -17,16 +24,19 @@ export default {
   methods: {
     action(type) {
       let parent = this.$parent.$parent;
+      parent.filter.title = '';
 
       if ( 'folder' === type ) {
-        this.$set(parent.filter, 'folder', parent.filter.folder + this.label + '/');
+        //this.$set(parent.filter, 'folder', parent.filter.folder + this.label + '/');
+        parent.filter['folder'] = parent.filter.folder + this.label + '/';
         parent.clearData();
         parent.get();
       } else if ( 'back' === type ) {
         let array = parent.filter.folder.split('/');
         array.pop();
         array.pop();
-        this.$set(parent.filter, 'folder', array.join('/') + '/');
+        //this.$set(parent.filter, 'folder', array.join('/') + '/');
+        parent.filter['folder'] = array.join('/') + '/';
         parent.clearData();
         parent.get();
       } else if ( 'remove' === type ) {
@@ -34,9 +44,10 @@ export default {
 
         Nova.request().post('/nova-vendor/nova-media-library/folder/del', { folder: parent.filter.folder }).then(r => {
           if ( r.data.folders )
-            this.$set(this.$parent.$parent.config, 'folders', r.data.folders);
+            //this.$set(this.$parent.$parent.config, 'folders', r.data.folders);
+            this.$parent.$parent.config['folders'] = r.data.folders;
           if ( r.data.message )
-            this.$toasted.show(r.data.message, { type: 'success' });
+            toasted.show(r.data.message, { type: 'success' });
           this.action('back');
         }).catch(e => {
           window.nmlToastHook(e);
@@ -50,9 +61,10 @@ export default {
           folder: folder
         }).then(r => {
           if ( r.data.folders )
-            this.$set(this.$parent.$parent.config, 'folders', r.data.folders);
+            //this.$set(this.$parent.$parent.config, 'folders', r.data.folders);
+            this.$parent.$parent.config['folders'] = r.data.folders;
           if ( r.data.message )
-            this.$toasted.show(r.data.message, { type: 'success' });
+            toasted.show(r.data.message, { type: 'success' });
         }).catch(e => {
           window.nmlToastHook(e);
         });
@@ -63,4 +75,5 @@ export default {
     this.title.folder = this.label;
     this.className = 'folder' === this.type ? '' : '-'+this.type;
   }
+  
 }
