@@ -1,9 +1,10 @@
 <template>
   <div :class="'flex flex-wrap nml-display-'+$parent.config.display">
+    
 
     <template v-if="'folders' === this.$parent.config.store">
       <template v-if="'/' !== folder">
-        <div class="bg-90 text-white rounded p-1 mx-1 mb-4 w-full break-words" style="font-family:monospace">{{ folder }}</div>
+        <div class="pl-3 text-white rounded p-1 mx-1 mb-4 w-full break-words bg-green-600 text-white font-bold" style="font-family:monospace">{{ folder }}</div>
         <Folders :key="folder" type="back" />
         <Folders :key="folder" v-if="!$parent.items.array.length && !getFolders.length" type="remove" />
       </template>
@@ -17,7 +18,7 @@
         v-for="item in $parent.items.array"
         :key="item.id"
         @click="clickItem(item)"
-        :class="['nml-item relative mb-2 cursor-pointer', { checked: $parent.bulk.enable && $parent.bulk.ids[item.id] }]"
+        :class="['nml-item relative mb-2 cursor-pointer dark:nml-bg-gray-800 bg-white', ($parent.bulk.enable && $parent.bulk.ids[item.id]) ? checked:'' ]"
         :title="item.title || item.name"
       >
         <div
