@@ -1,5 +1,10 @@
+import Datepicker from 'vue3-datepicker'
+import { ref } from 'vue'
+
 export default {
+  components: { Datepicker },
   data() {
+    let picked = ref(new Date());
     let all = null;
     let types = this.$parent.field ? (this.$parent.$props.types || []) : [];
     if ( !types.length ) {
@@ -9,13 +14,28 @@ export default {
     }
     return {
       all,
-      types
+      types,
+      picked,
+      uploadFromText:'Upload From',
+      uploadToText:'Upload To',
+      uploadFrom:new Date(),
+      uploadTo:new Date()
+      
+    }
+  },
+  watch: {
+    uploadFrom () {
+      return this.updateDate(this.$refs.uploadFrom.input, 'from');
+    },
+    uploadTo () {
+      return this.updateDate(this.$refs.uploadTo.input, 'to');
     }
   },
   methods: {
     updateDate(value, target) {
       this.$parent.filter[target] = value || null;
       this.$parent.doSearch();
+      
     },
     display(val) {
       this.$parent.config.display = val;
