@@ -8,24 +8,24 @@ use Illuminate\Http\Exceptions\HttpResponseException;
 
 class UploadFr extends FormRequest
 {
-	protected function failedValidation(Validator $validator) {
-		throw new HttpResponseException(response()->json([
-			'message' => $validator->errors()->first()
-		], 422));
-	}
+    protected function failedValidation(Validator $validator)
+    {
+        throw new HttpResponseException(response()->json([
+            'message' => $validator->errors()->first(),
+        ], 422));
+    }
 
-	public function rules()
-	{
-		return [
-			'file' => 'required|file'
-		];
-	}
+    public function rules()
+    {
+        return [
+            'file' => 'required|file',
+        ];
+    }
 
-	public function messages()
-	{
-		return [
-			'file.*' => __('Invalid file')
-		];
-	}
-
+    public function messages()
+    {
+        return [
+            'file.*' => __('Invalid file'),
+        ];
+    }
 }
