@@ -2,6 +2,7 @@
 
 namespace ClassicO\NovaMediaLibrary;
 
+use Laravel\Nova\Nova;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 use ClassicO\NovaMediaLibrary\Http\Middleware\Authorize;
@@ -15,15 +16,15 @@ class ToolServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        $this->loadViewsFrom(__DIR__.'/../resources/views', 'nova-media-library');
+        $this->loadViewsFrom(__DIR__ . '/../resources/views', 'nova-media-library');
         $this->loadJsonTranslationsFrom(resource_path('lang/vendor/nova-media-library'));
-        $this->loadMigrationsFrom(__DIR__.'/../database/');
+        $this->loadMigrationsFrom(__DIR__ . '/../database/');
 
-	    $this->publishes([
-		    __DIR__.'/../config/' => config_path(),
-		    __DIR__.'/../database/' => base_path('/database/migrations'),
-		    __DIR__.'/../resources/lang' => resource_path('lang/vendor/nova-media-library'),
-	    ], 'config-nml');
+        $this->publishes([
+            __DIR__ . '/../config/' => config_path(),
+            __DIR__ . '/../database/' => base_path('/database/migrations'),
+            __DIR__ . '/../resources/lang' => resource_path('lang/vendor/nova-media-library'),
+        ], 'config-nml');
 
         $this->app->booted(function () {
             $this->routes();
@@ -41,9 +42,12 @@ class ToolServiceProvider extends ServiceProvider
             return;
         }
 
+        Nova::router(['nova', Authorize::class], 'nova-media-library')
+            ->group(__DIR__ . '/../routes/inertia.php');
+
         Route::middleware(['nova', Authorize::class])
-                ->prefix('nova-vendor/nova-media-library')
-                ->group(__DIR__.'/../routes/api.php');
+            ->prefix('nova-vendor/nova-media-library')
+            ->group(__DIR__ . '/../routes/api.php');
     }
 
     /**
@@ -53,6 +57,5 @@ class ToolServiceProvider extends ServiceProvider
      */
     public function register()
     {
-
     }
 }
